test_requirements.txt
fabric3
mixer
mock
coverage
django-coverage
ipdb
flake8
tox
